[ETCM-355] Move ProtocolFamily next to Capability
