+++ src/main/scala/io/iohk/ethereum/network/p2p/messages/package.scala
@@ -1,23 +1,5 @@
 package io.iohk.ethereum.network.p2p
 
-import io.iohk.ethereum.rlp.RLPEncodeable
-import io.iohk.ethereum.rlp._
-
-package messages {
-  sealed trait ProtocolFamily
-  object ProtocolFamily {
-    final case object ETH extends ProtocolFamily
-    final case object ETC extends ProtocolFamily
-
-    implicit class ProtocolFamilyEnc(val msg: ProtocolFamily) extends RLPSerializable {
-      override def toRLPEncodable: RLPEncodeable = msg match {
-        case ETH => RLPValue("eth".getBytes())
-        case ETC => RLPValue("etc".getBytes())
-      }
-    }
-  }
-}
-
 package object messages {
 
   val SubProtocolOffset = 0x10
